# docs/api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "GOAT AI",
    "description": "Simon Business School Strategic Intelligence API",
    "version": "1.2.0"
  },
  "paths": {
...
    "/api/system/gpu": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Read GPU telemetry",
        "description": "Return real-time GPU telemetry used by frontend status strip.",
        "operationId": "get_gpu_status_api_system_gpu_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GPUStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/inference": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Read rolling chat latency telemetry",
        "description": "Rolling average duration of completed chat streams (milliseconds).",
        "operationId": "get_inference_latency_route_api_system_inference_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceLatencyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/features": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Read capability-gated feature flags",
        "description": "Return config + host probes for optional high-risk features (see docs/standards/ENGINEERING_STANDARDS.md §15).",
        "operationId": "get_system_features_route_api_system_features_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemFeaturesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/runtime-target": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Read resolved deployment target order",
        "description": "Return the current runtime target and the ordered deployment fallback list.",
        "operationId": "get_runtime_target_route_api_system_runtime_target_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeTargetResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodeSandboxFeaturePayload": {
        "properties": {
          "policy_allowed": {
            "type": "boolean",
            "title": "Policy Allowed",
            "description": "Caller-scoped policy gate derived from credential scopes."
          },
          "allowed_by_config": {
            "type": "boolean",
            "title": "Allowed By Config",
            "description": "Operator enabled the feature in deployment config."
          },
          "available_on_host": {
            "type": "boolean",
            "title": "Available On Host",
            "description": "Probed runtime dependency (e.g. Docker) appears usable."
          },
          "effective_enabled": {
            "type": "boolean",
            "title": "Effective Enabled",
            "description": "Both config and probe satisfied."
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name",
            "description": "Selected runtime provider for code sandbox execution."
          },
          "isolation_level": {
            "type": "string",
            "title": "Isolation Level",
            "description": "Isolation tier for the active provider: container or host."
          },
          "network_policy_enforced": {
            "type": "boolean",
            "title": "Network Policy Enforced",
            "description": "Whether the selected provider can actually enforce the declared network policy."
          },
          "deny_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deny Reason",
            "description": "Controlled enum for runtime gate only; never raw exception text."
          }
        },
        "type": "object",
        "required": [
          "policy_allowed",
          "allowed_by_config",
          "available_on_host",
          "effective_enabled",
          "provider_name",
          "isolation_level",
          "network_policy_enforced"
        ],
        "title": "CodeSandboxFeaturePayload",
        "description": "Runtime / dependency readiness for the code sandbox (§15).\n\n``policy_allowed`` is evaluated from the current request's authorization context.\nRuntime availability remains separate and must not be conflated with policy denial."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Detail",
            "description": "Human-readable message or structured validation errors."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Echo of X-Request-ID when the server generated or accepted one."
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "Standard JSON error envelope (Phase 13 §13.0)."
      },
      "GPUStatusResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "uuid": {
            "type": "string",
            "title": "Uuid",
            "default": ""
          },
          "utilization_gpu": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utilization Gpu"
          },
          "memory_used_mb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Used Mb"
          },
          "memory_total_mb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Total Mb"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "power_draw_w": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Draw W"
          }
        },
        "type": "object",
        "required": [
          "available",
          "active",
          "message"
        ],
        "title": "GPUStatusResponse",
        "description": "GPU telemetry payload used by sidebar status strip."
      },
...
      "InferenceLatencyResponse": {
        "properties": {
          "chat_avg_ms": {
            "type": "number",
            "title": "Chat Avg Ms"
          },
          "chat_sample_count": {
            "type": "integer",
            "title": "Chat Sample Count"
          },
          "chat_p50_ms": {
            "type": "number",
            "title": "Chat P50 Ms"
          },
          "chat_p95_ms": {
            "type": "number",
            "title": "Chat P95 Ms"
          },
          "first_token_avg_ms": {
            "type": "number",
            "title": "First Token Avg Ms"
          },
          "first_token_sample_count": {
            "type": "integer",
            "title": "First Token Sample Count"
          },
          "first_token_p50_ms": {
            "type": "number",
            "title": "First Token P50 Ms"
          },
          "first_token_p95_ms": {
            "type": "number",
            "title": "First Token P95 Ms"
          },
          "model_buckets": {
            "additionalProperties": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "object"
            },
            "type": "object",
            "title": "Model Buckets"
          }
        },
        "type": "object",
        "required": [
          "chat_avg_ms",
          "chat_sample_count",
          "chat_p50_ms",
          "chat_p95_ms",
          "first_token_avg_ms",
          "first_token_sample_count",
          "first_token_p50_ms",
          "first_token_p95_ms",
          "model_buckets"
        ],
        "title": "InferenceLatencyResponse",
        "description": "Rolling average chat completion time (server-side, in-process)."
      },
...
      "RuntimeFeaturePayload": {
        "properties": {
          "allowed_by_config": {
            "type": "boolean",
            "title": "Allowed By Config",
            "description": "Operator enabled the feature family in deployment config."
          },
          "available_on_host": {
            "type": "boolean",
            "title": "Available On Host",
            "description": "Required runtime dependencies and executors are available."
          },
          "effective_enabled": {
            "type": "boolean",
            "title": "Effective Enabled",
            "description": "Both config and runtime are ready."
          },
          "deny_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deny Reason",
            "description": "Controlled enum for runtime gate only; never raw exception text."
          }
        },
        "type": "object",
        "required": [
          "allowed_by_config",
          "available_on_host",
          "effective_enabled"
        ],
        "title": "RuntimeFeaturePayload",
        "description": "Runtime / dependency readiness for optional non-policy-gated features."
      },
      "RuntimeOperationalContractResponse": {
        "properties": {
          "storage_model": {
            "type": "string",
            "title": "Storage Model"
          },
          "concurrency_model": {
            "type": "string",
            "title": "Concurrency Model"
          },
          "process_local_seams": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Process Local Seams"
          },
          "scaling_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scaling Notes"
          }
        },
        "type": "object",
        "required": [
          "storage_model",
          "concurrency_model",
          "process_local_seams",
          "scaling_notes"
        ],
        "title": "RuntimeOperationalContractResponse",
        "description": "Operator-facing runtime contract for scaling and recovery expectations."
      },
      "RuntimeTargetItemResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "host": {
            "type": "string",
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "host",
          "port",
          "base_url",
          "reason"
        ],
        "title": "RuntimeTargetItemResponse",
        "description": "One runtime target candidate in deployment preference order."
      },
      "RuntimeTargetResponse": {
        "properties": {
          "deploy_target": {
            "type": "string",
            "title": "Deploy Target"
          },
          "current": {
            "$ref": "#/components/schemas/RuntimeTargetItemResponse"
          },
          "ordered_targets": {
            "items": {
              "$ref": "#/components/schemas/RuntimeTargetItemResponse"
            },
            "type": "array",
            "title": "Ordered Targets"
          },
          "operational_contract": {
            "$ref": "#/components/schemas/RuntimeOperationalContractResponse"
          }
        },
        "type": "object",
        "required": [
          "deploy_target",
          "current",
          "ordered_targets",
          "operational_contract"
        ],
        "title": "RuntimeTargetResponse",
        "description": "Resolved runtime target metadata used by deploy/ops tooling."
      },
      "SystemFeaturesResponse": {
        "properties": {
          "code_sandbox": {
            "$ref": "#/components/schemas/CodeSandboxFeaturePayload"
          },
          "workbench": {
            "$ref": "#/components/schemas/WorkbenchFeaturePayload"
          }
        },
        "type": "object",
        "required": [
          "code_sandbox",
          "workbench"
        ],
        "title": "SystemFeaturesResponse",
        "description": "Public feature flags derived from config + host probes."
      },
...
      "WorkbenchFeaturePayload": {
        "properties": {
          "agent_tasks": {
            "$ref": "#/components/schemas/RuntimeFeaturePayload"
          },
          "plan_mode": {
            "$ref": "#/components/schemas/RuntimeFeaturePayload"
          },
          "browse": {
            "$ref": "#/components/schemas/RuntimeFeaturePayload"
          },
          "deep_research": {
            "$ref": "#/components/schemas/RuntimeFeaturePayload"
          },
          "artifact_workspace": {
            "$ref": "#/components/schemas/RuntimeFeaturePayload"
          },
          "project_memory": {
            "$ref": "#/components/schemas/RuntimeFeaturePayload"
          },
          "connectors": {
            "$ref": "#/components/schemas/RuntimeFeaturePayload"
          }
        },
        "type": "object",
        "required": [
          "agent_tasks",
          "plan_mode",
          "browse",
          "deep_research",
          "artifact_workspace",
          "project_memory",
          "connectors"
        ],
        "title": "WorkbenchFeaturePayload",
        "description": "Capability inventory for future agent/workbench surfaces."
      },
...
    }
  },
...
}